Added cmd("setStepLongLong")
In order to store step as long long, I had to change all step-like
variables throughout the whole code.

Notice that on Linux/MacOS this should have no effects
(long = long long)
However, on Windows it will promote step to 64 bits.

+++ src/generic/EffectiveEnergyDrift.cpp
@@ -74,7 +74,7 @@ EFFECTIVE_ENERGY_DRIFT PRINT_STRIDE=100 FILE=eff
 class EffectiveEnergyDrift:
   public ActionPilot {
   OFile output;
-  long int printStride;
+  long long int printStride;
   std::string fmt;
 
   double eed;
@@ -82,7 +82,7 @@ class EffectiveEnergyDrift:
   Atoms& atoms;
   std::vector<ActionWithValue*> biases;
 
-  long int pDdStep;
+  long long int pDdStep;
   int nLocalAtoms;
   int pNLocalAtoms;
   std::vector<int> pGatindex;
