+++ src/opes/OPESexpanded.cpp
@@ -120,7 +120,7 @@ private:
   unsigned rank_;
   unsigned NumWalkers_;
   unsigned walker_rank_;
-  unsigned long counter_;
+  unsigned long long counter_;
   std::size_t ncv_;
 
   std::vector<const double *> ECVs_;
